use unified tui for dagger do
