+++ cmd/dagger/query.go
@@ -8,7 +8,6 @@ import (
 	"os"
 	"strings"
 
-	"github.com/dagger/dagger/internal/engine/journal"
 	"github.com/dagger/dagger/router"
 	"github.com/spf13/cobra"
 	"golang.org/x/term"
@@ -91,7 +90,7 @@ func Query(cmd *cobra.Command, args []string) {
 
 func doQuery(ctx context.Context, query, op string, vars map[string]interface{}) ([]byte, error) {
 	res := make(map[string]interface{})
-	err := withEngine(ctx, "", journal.Discard{}, os.Stderr, func(ctx context.Context, r *router.Router) error {
+	err := withEngineAndTUI(ctx, func(ctx context.Context, r *router.Router, sessionToken string) error {
 		_, err := r.Do(ctx, query, op, vars, &res)
 		return err
 	})
